Refactored cron service, implemented new unit tests
