+++ packages/apps/reputation-oracle/server/src/modules/cron-job/cron-job.module.ts
@@ -4,34 +4,17 @@ import { CronJobService } from './cron-job.service';
 import { CronJobRepository } from './cron-job.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CronJobEntity } from './cron-job.entity';
-import { Web3Module } from '../web3/web3.module';
 import { WebhookModule } from '../webhook/webhook.module';
 import { ConfigModule } from '@nestjs/config';
-import { PayoutModule } from '../payout/payout.module';
-import { ReputationModule } from '../reputation/reputation.module';
-import { WebhookIncomingRepository } from '../webhook/webhook-incoming.repository';
-import { WebhookOutgoingRepository } from '../webhook/webhook-outgoing.repository';
-import { EscrowCompletionTrackingRepository } from '../escrow-completion-tracking/escrow-completion-tracking.repository';
-import { EscrowCompletionTrackingModule } from '../escrow-completion-tracking/escrow-completion-tracking.module';
 
 @Global()
 @Module({
   imports: [
     TypeOrmModule.forFeature([CronJobEntity]),
     ConfigModule,
-    Web3Module,
     WebhookModule,
-    EscrowCompletionTrackingModule,
-    PayoutModule,
-    ReputationModule,
   ],
-  providers: [
-    CronJobService,
-    CronJobRepository,
-    WebhookIncomingRepository,
-    WebhookOutgoingRepository,
-    EscrowCompletionTrackingRepository,
-  ],
+  providers: [CronJobService, CronJobRepository],
   exports: [CronJobService],
 })
 export class CronJobModule {}
