+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook.module.ts
@@ -10,13 +10,24 @@ import { WebhookIncomingEntity } from './webhook-incoming.entity';
 import { WebhookOutgoingEntity } from './webhook-outgoing.entity';
 import { WebhookIncomingRepository } from './webhook-incoming.repository';
 import { WebhookOutgoingRepository } from './webhook-outgoing.repository';
+import { EscrowCompletionTrackingModule } from '../escrow-completion-tracking/escrow-completion-tracking.module';
+import { EscrowCompletionTrackingRepository } from '../escrow-completion-tracking/escrow-completion-tracking.repository';
+import { PayoutModule } from '../payout/payout.module';
+import { ReputationModule } from '../reputation/reputation.module';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([WebhookIncomingEntity, WebhookOutgoingEntity]),
+    TypeOrmModule.forFeature([
+      WebhookIncomingEntity,
+      WebhookOutgoingEntity,
+      EscrowCompletionTrackingRepository,
+    ]),
     ConfigModule,
     Web3Module,
     HttpModule,
+    EscrowCompletionTrackingModule,
+    PayoutModule,
+    ReputationModule,
   ],
   controllers: [WebhookController],
   providers: [
