+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook.service.spec.ts
@@ -1,5 +1,5 @@
 import { createMock } from '@golevelup/ts-jest';
-import { ChainId } from '@human-protocol/sdk';
+import { ChainId, EscrowClient, EscrowStatus } from '@human-protocol/sdk';
 import { HttpService } from '@nestjs/axios';
 import { ConfigService } from '@nestjs/config';
 import { Test } from '@nestjs/testing';
@@ -13,6 +13,7 @@ import {
   mockConfig,
 } from '../../../test/constants';
 import {
+  EscrowCompletionTrackingStatus,
   EventType,
   WebhookIncomingStatus,
   WebhookOutgoingStatus,
@@ -32,50 +33,96 @@ import { HttpStatus } from '@nestjs/common';
 import { Web3ConfigService } from '../../common/config/web3-config.service';
 import { ServerConfigService } from '../../common/config/server-config.service';
 import { ControlledError } from '../../common/errors/controlled';
+import { PayoutService } from '../payout/payout.service';
+import { ReputationService } from '../reputation/reputation.service';
+import { EscrowCompletionTrackingRepository } from '../escrow-completion-tracking/escrow-completion-tracking.repository';
+import { EscrowCompletionTrackingService } from '../escrow-completion-tracking/escrow-completion-tracking.service';
+import { ReputationConfigService } from '../../common/config/reputation-config.service';
+import { ReputationRepository } from '../reputation/reputation.repository';
+import { StorageService } from '../storage/storage.service';
+import { PostgresErrorCodes } from '../../common/enums/database';
+import { DatabaseError } from '../../common/errors/database';
 
 jest.mock('@human-protocol/sdk', () => ({
   ...jest.requireActual('@human-protocol/sdk'),
   EscrowClient: {
-    build: jest.fn(),
+    build: jest.fn().mockImplementation(() => ({
+      createEscrow: jest.fn().mockResolvedValue(MOCK_ADDRESS),
+      getJobLauncherAddress: jest.fn().mockResolvedValue(MOCK_ADDRESS),
+      getExchangeOracleAddress: jest.fn().mockResolvedValue(MOCK_ADDRESS),
+      getRecordingOracleAddress: jest.fn().mockResolvedValue(MOCK_ADDRESS),
+      setup: jest.fn().mockResolvedValue(null),
+      fund: jest.fn().mockResolvedValue(null),
+      getManifestUrl: jest.fn().mockResolvedValue(MOCK_FILE_URL),
+      complete: jest.fn().mockResolvedValue(null),
+      getStatus: jest.fn(),
+    })),
   },
-  KVStoreClient: {
-    build: jest.fn(),
+  KVStoreUtils: {
+    get: jest.fn(),
+  },
+  OperatorUtils: {
+    getLeader: jest.fn().mockImplementation(() => {
+      return { webhookUrl: MOCK_WEBHOOK_URL };
+    }),
   },
 }));
 
 describe('WebhookService', () => {
-  let webhookService: WebhookService,
-    webhookIncomingRepository: WebhookIncomingRepository,
-    webhookOutgoingRepository: WebhookOutgoingRepository,
-    httpService: HttpService,
-    web3ConfigService: Web3ConfigService;
+  let webhookService: WebhookService;
+  let webhookIncomingRepository: WebhookIncomingRepository;
+  let webhookOutgoingRepository: WebhookOutgoingRepository;
+  let httpService: HttpService;
+  let web3ConfigService: Web3ConfigService;
+  let escrowCompletionTrackingService: EscrowCompletionTrackingService;
+  let escrowCompletionTrackingRepository: EscrowCompletionTrackingRepository;
+  let payoutService: PayoutService;
+  let reputationService: ReputationService;
 
   const signerMock = {
     address: MOCK_ADDRESS,
     getNetwork: jest.fn().mockResolvedValue({ chainId: 1 }),
   };
 
+  // Mock ConfigService to return the mock configuration values
+  const mockConfigService = {
+    get: jest.fn((key: string) => mockConfig[key]),
+    getOrThrow: jest.fn((key: string) => {
+      if (!mockConfig[key])
+        throw new Error(`Configuration key "${key}" does not exist`);
+      return mockConfig[key];
+    }),
+  };
+
+  // Mock Web3Service
+  const mockWeb3Service = {
+    getSigner: jest.fn().mockReturnValue(signerMock),
+    validateChainId: jest.fn().mockReturnValue(new Error()),
+    calculateGasPrice: jest.fn().mockReturnValue(1000n),
+    getOperatorAddress: jest.fn().mockReturnValue(MOCK_ADDRESS),
+  };
+
   beforeEach(async () => {
     const moduleRef = await Test.createTestingModule({
       providers: [
+        WebhookService,
+        EscrowCompletionTrackingService,
+        PayoutService,
+        ReputationService,
+        Web3ConfigService,
+        ServerConfigService,
+        ReputationConfigService,
+        HttpService,
+        StorageService,
+
+        // Mocked services
         {
           provide: ConfigService,
-          useValue: {
-            get: jest.fn((key: string) => mockConfig[key]),
-            getOrThrow: jest.fn((key: string) => {
-              if (!mockConfig[key]) {
-                throw new Error(`Configuration key "${key}" does not exist`);
-              }
-              return mockConfig[key];
-            }),
-          },
+          useValue: mockConfigService,
         },
-        WebhookService,
         {
           provide: Web3Service,
-          useValue: {
-            getSigner: jest.fn().mockReturnValue(signerMock),
-          },
+          useValue: mockWeb3Service,
         },
         {
           provide: WebhookIncomingRepository,
@@ -85,18 +132,36 @@ describe('WebhookService', () => {
           provide: WebhookOutgoingRepository,
           useValue: createMock<WebhookOutgoingRepository>(),
         },
-        Web3ConfigService,
-        ServerConfigService,
+        {
+          provide: EscrowCompletionTrackingRepository,
+          useValue: createMock<EscrowCompletionTrackingRepository>(),
+        },
+        {
+          provide: ReputationRepository,
+          useValue: createMock<ReputationRepository>(),
+        },
         { provide: HttpService, useValue: createMock<HttpService>() },
+        { provide: StorageService, useValue: createMock<StorageService>() },
       ],
     }).compile();
 
+    // Assign injected dependencies to variables
     webhookService = moduleRef.get<WebhookService>(WebhookService);
     webhookIncomingRepository = moduleRef.get(WebhookIncomingRepository);
     webhookOutgoingRepository = moduleRef.get(WebhookOutgoingRepository);
+    escrowCompletionTrackingRepository = moduleRef.get(
+      EscrowCompletionTrackingRepository,
+    );
+    payoutService = moduleRef.get<PayoutService>(PayoutService);
+    reputationService = moduleRef.get<ReputationService>(ReputationService);
+    escrowCompletionTrackingService =
+      moduleRef.get<EscrowCompletionTrackingService>(
+        EscrowCompletionTrackingService,
+      );
     httpService = moduleRef.get(HttpService);
     web3ConfigService = moduleRef.get(Web3ConfigService);
 
+    // Mocking privateKey getter
     jest
       .spyOn(web3ConfigService, 'privateKey', 'get')
       .mockReturnValue(MOCK_PRIVATE_KEY);
@@ -107,6 +172,18 @@ describe('WebhookService', () => {
   });
 
   describe('createIncomingWebhook', () => {
+    const validDto: IncomingWebhookDto = {
+      chainId: ChainId.LOCALHOST,
+      escrowAddress: MOCK_ADDRESS,
+      eventType: EventType.JOB_COMPLETED,
+    };
+
+    const invalidDto: IncomingWebhookDto = {
+      chainId: ChainId.LOCALHOST,
+      escrowAddress: MOCK_ADDRESS,
+      eventType: 'JOB_FAILED' as EventType,
+    };
+
     const webhookEntity: Partial<WebhookIncomingEntity> = {
       chainId: ChainId.LOCALHOST,
       escrowAddress: MOCK_ADDRESS,
@@ -115,13 +192,7 @@ describe('WebhookService', () => {
       retriesCount: 0,
     };
 
-    it('should successfully create incoming webhook with valid DTO', async () => {
-      const validDto: IncomingWebhookDto = {
-        chainId: ChainId.LOCALHOST,
-        escrowAddress: MOCK_ADDRESS,
-        eventType: EventType.JOB_COMPLETED,
-      };
-
+    it('should successfully create an incoming webhook with valid DTO', async () => {
       jest
         .spyOn(webhookIncomingRepository, 'createUnique')
         .mockResolvedValue(webhookEntity as WebhookIncomingEntity);
@@ -129,17 +200,14 @@ describe('WebhookService', () => {
       await webhookService.createIncomingWebhook(validDto);
 
       expect(webhookIncomingRepository.createUnique).toHaveBeenCalledWith(
-        expect.any(Object),
+        expect.objectContaining({
+          chainId: validDto.chainId,
+          escrowAddress: validDto.escrowAddress,
+        }),
       );
     });
 
-    it('should throw NotFoundException if webhook entity not created', async () => {
-      const validDto: IncomingWebhookDto = {
-        chainId: ChainId.LOCALHOST,
-        escrowAddress: MOCK_ADDRESS,
-        eventType: EventType.JOB_COMPLETED,
-      };
-
+    it('should throw NotFoundException if webhook entity creation fails', async () => {
       jest
         .spyOn(webhookIncomingRepository as any, 'createUnique')
         .mockResolvedValue(null);
@@ -151,13 +219,7 @@ describe('WebhookService', () => {
       );
     });
 
-    it('should throw BadRequestException with invalid event type', async () => {
-      const invalidDto: IncomingWebhookDto = {
-        chainId: ChainId.LOCALHOST,
-        escrowAddress: MOCK_ADDRESS,
-        eventType: 'JOB_FAILED' as EventType,
-      };
-
+    it('should throw BadRequestException with an invalid event type', async () => {
       await expect(
         webhookService.createIncomingWebhook(invalidDto),
       ).rejects.toThrow(
@@ -202,31 +264,74 @@ describe('WebhookService', () => {
   });
 
   describe('handleWebhookIncomingError', () => {
-    it('should set incoming webhook status to FAILED if retries exceed threshold', async () => {
-      const webhookEntity: Partial<WebhookIncomingEntity> = {
-        id: 1,
-        status: WebhookIncomingStatus.PENDING,
-        retriesCount: MOCK_MAX_RETRY_COUNT,
-      };
-      await (webhookService as any).handleWebhookIncomingError(
-        webhookEntity,
-        new Error('Sample error'),
-      );
-      expect(webhookIncomingRepository.updateOne).toHaveBeenCalled();
-      expect(webhookEntity.status).toBe(WebhookIncomingStatus.FAILED);
-    });
+    let webhookEntity: Partial<WebhookIncomingEntity>;
 
-    it('should increment retries count if below threshold', async () => {
-      const webhookEntity: Partial<WebhookIncomingEntity> = {
+    beforeEach(() => {
+      webhookEntity = {
         id: 1,
         status: WebhookIncomingStatus.PENDING,
         retriesCount: 0,
       };
+    });
+
+    it('should set incoming webhook status to FAILED if retries exceed threshold', async () => {
+      webhookEntity.retriesCount = MOCK_MAX_RETRY_COUNT;
+
       await (webhookService as any).handleWebhookIncomingError(
         webhookEntity,
         new Error('Sample error'),
       );
-      expect(webhookIncomingRepository.updateOne).toHaveBeenCalled();
+
+      expect(webhookIncomingRepository.updateOne).toHaveBeenCalledWith(
+        expect.objectContaining({
+          id: webhookEntity.id,
+          status: WebhookIncomingStatus.FAILED,
+        }),
+      );
+      expect(webhookEntity.status).toBe(WebhookIncomingStatus.FAILED);
+    });
+
+    it('should increment retries count if below threshold', async () => {
+      webhookEntity.retriesCount = 0;
+
+      await (webhookService as any).handleWebhookIncomingError(
+        webhookEntity,
+        new Error('Sample error'),
+      );
+
+      expect(webhookIncomingRepository.updateOne).toHaveBeenCalledWith(
+        expect.objectContaining({
+          id: webhookEntity.id,
+          status: WebhookIncomingStatus.PENDING,
+          retriesCount: 1,
+        }),
+      );
+      expect(webhookEntity.status).toBe(WebhookIncomingStatus.PENDING);
+      expect(webhookEntity.retriesCount).toBe(1);
+      expect(webhookEntity.waitUntil).toBeInstanceOf(Date);
+    });
+
+    it('should throw an error if the update operation fails', async () => {
+      jest
+        .spyOn(webhookIncomingRepository, 'updateOne')
+        .mockRejectedValue(new Error('Database error'));
+
+      await expect(
+        (webhookService as any).handleWebhookIncomingError(
+          webhookEntity,
+          new Error('Sample error'),
+        ),
+      ).rejects.toThrow('Database error');
+    });
+
+    it('should handle webhook error gracefully when retries count is zero', async () => {
+      webhookEntity.retriesCount = 0;
+
+      await (webhookService as any).handleWebhookIncomingError(
+        webhookEntity,
+        new Error('Network failure'),
+      );
+
       expect(webhookEntity.status).toBe(WebhookIncomingStatus.PENDING);
       expect(webhookEntity.retriesCount).toBe(1);
       expect(webhookEntity.waitUntil).toBeInstanceOf(Date);
@@ -234,31 +339,67 @@ describe('WebhookService', () => {
   });
 
   describe('handleWebhookOutgoingError', () => {
-    it('should set outgoing webhook status to FAILED if retries exceed threshold', async () => {
-      const webhookEntity: Partial<WebhookOutgoingEntity> = {
-        id: 1,
-        status: WebhookOutgoingStatus.PENDING,
-        retriesCount: MOCK_MAX_RETRY_COUNT,
-      };
-      await (webhookService as any).handleWebhookOutgoingError(
-        webhookEntity,
-        new Error('Sample error'),
-      );
-      expect(webhookOutgoingRepository.updateOne).toHaveBeenCalled();
-      expect(webhookEntity.status).toBe(WebhookOutgoingStatus.FAILED);
-    });
+    let webhookEntity: Partial<WebhookOutgoingEntity>;
 
-    it('should increment retries count if below threshold', async () => {
-      const webhookEntity: Partial<WebhookOutgoingEntity> = {
+    beforeEach(() => {
+      webhookEntity = {
         id: 1,
         status: WebhookOutgoingStatus.PENDING,
         retriesCount: 0,
       };
+    });
+
+    it('should set outgoing webhook status to FAILED if retries exceed threshold', async () => {
+      webhookEntity.retriesCount = MOCK_MAX_RETRY_COUNT;
       await (webhookService as any).handleWebhookOutgoingError(
         webhookEntity,
         new Error('Sample error'),
       );
-      expect(webhookOutgoingRepository.updateOne).toHaveBeenCalled();
+      expect(webhookOutgoingRepository.updateOne).toHaveBeenCalledWith(
+        expect.objectContaining({
+          id: webhookEntity.id,
+          status: WebhookOutgoingStatus.FAILED,
+        }),
+      );
+      expect(webhookEntity.status).toBe(WebhookOutgoingStatus.FAILED);
+    });
+
+    it('should increment retries count and set waitUntil if retries are below threshold', async () => {
+      webhookEntity.retriesCount = 0;
+      await (webhookService as any).handleWebhookOutgoingError(
+        webhookEntity,
+        new Error('Sample error'),
+      );
+      expect(webhookOutgoingRepository.updateOne).toHaveBeenCalledWith(
+        expect.objectContaining({
+          id: webhookEntity.id,
+          status: WebhookOutgoingStatus.PENDING,
+          retriesCount: 1,
+        }),
+      );
+      expect(webhookEntity.status).toBe(WebhookOutgoingStatus.PENDING);
+      expect(webhookEntity.retriesCount).toBe(1);
+      expect(webhookEntity.waitUntil).toBeInstanceOf(Date);
+    });
+
+    it('should throw an error if the update operation fails', async () => {
+      jest
+        .spyOn(webhookOutgoingRepository, 'updateOne')
+        .mockRejectedValue(new Error('Database error'));
+      await expect(
+        (webhookService as any).handleWebhookOutgoingError(
+          webhookEntity,
+          new Error('Sample error'),
+        ),
+      ).rejects.toThrow('Database error');
+    });
+
+    it('should handle webhook outgoing error gracefully when retries count is zero', async () => {
+      webhookEntity.retriesCount = 0;
+      await (webhookService as any).handleWebhookOutgoingError(
+        webhookEntity,
+        new Error('Network failure'),
+      );
       expect(webhookEntity.status).toBe(WebhookOutgoingStatus.PENDING);
       expect(webhookEntity.retriesCount).toBe(1);
       expect(webhookEntity.waitUntil).toBeInstanceOf(Date);
@@ -310,6 +451,328 @@ describe('WebhookService', () => {
       ).rejects.toThrow(
         new ControlledError(ErrorWebhook.NotSent, HttpStatus.BAD_REQUEST),
       );
+    });
+  });
+
+  describe('processPendingIncomingWebhooks', () => {
+    let createEscrowCompletionTrackingMock: any;
+    let webhookEntity1: Partial<WebhookIncomingEntity>,
+      webhookEntity2: Partial<WebhookIncomingEntity>;
+
+    beforeEach(() => {
+      webhookEntity1 = {
+        id: 1,
+        chainId: ChainId.LOCALHOST,
+        escrowAddress: MOCK_ADDRESS,
+        status: WebhookIncomingStatus.PENDING,
+        waitUntil: new Date(),
+        retriesCount: 0,
+      };
+
+      webhookEntity2 = {
+        id: 2,
+        chainId: ChainId.LOCALHOST,
+        escrowAddress: MOCK_ADDRESS,
+        status: WebhookIncomingStatus.PENDING,
+        waitUntil: new Date(),
+        retriesCount: 0,
+      };
+
+      jest
+        .spyOn(webhookIncomingRepository, 'findByStatus')
+        .mockResolvedValue([webhookEntity1 as any, webhookEntity2 as any]);
+
+      createEscrowCompletionTrackingMock = jest.spyOn(
+        escrowCompletionTrackingService as any,
+        'createEscrowCompletionTracking',
+      );
+      createEscrowCompletionTrackingMock.mockResolvedValue(undefined);
+    });
+
+    afterEach(() => {
+      jest.restoreAllMocks();
+    });
+
+    it('should send webhook for all of the pending incoming webhooks', async () => {
+      await webhookService.processPendingIncomingWebhooks();
+
+      expect(createEscrowCompletionTrackingMock).toHaveBeenCalledTimes(2);
+      expect(createEscrowCompletionTrackingMock).toHaveBeenCalledWith(
+        webhookEntity1.chainId,
+        webhookEntity1.escrowAddress,
+      );
+      expect(createEscrowCompletionTrackingMock).toHaveBeenCalledWith(
+        webhookEntity2.chainId,
+        webhookEntity2.escrowAddress,
+      );
+
+      expect(webhookIncomingRepository.updateOne).toHaveBeenCalledTimes(2);
+      expect(webhookEntity1.status).toBe(WebhookIncomingStatus.COMPLETED);
+      expect(webhookEntity2.status).toBe(WebhookIncomingStatus.COMPLETED);
+    });
+
+    it('should increase retriesCount by 1 if sending webhook fails', async () => {
+      createEscrowCompletionTrackingMock.mockRejectedValueOnce(new Error());
+      await webhookService.processPendingIncomingWebhooks();
+
+      expect(webhookIncomingRepository.updateOne).toHaveBeenCalled();
+      expect(webhookEntity1.status).toBe(WebhookIncomingStatus.PENDING);
+      expect(webhookEntity1.retriesCount).toBe(1);
+      expect(webhookEntity1.waitUntil).toBeInstanceOf(Date);
+    });
+
+    it('should mark webhook as failed if retriesCount exceeds threshold', async () => {
+      createEscrowCompletionTrackingMock.mockRejectedValueOnce(new Error());
+
+      webhookEntity1.retriesCount = MOCK_MAX_RETRY_COUNT;
+
+      await webhookService.processPendingIncomingWebhooks();
+
+      expect(webhookIncomingRepository.updateOne).toHaveBeenCalled();
+      expect(webhookEntity1.status).toBe(WebhookIncomingStatus.FAILED);
+    });
+
+    it('should retry if createEscrowCompletionTracking throws a DatabaseError with Duplicated error code', async () => {
+      const mockEntity = { id: 1, status: EscrowCompletionTrackingStatus.PAID };
+
+      jest
+        .spyOn(escrowCompletionTrackingRepository, 'findByStatus')
+        .mockResolvedValue([mockEntity as any]);
+
+      const updateOneMock = jest
+        .spyOn(escrowCompletionTrackingRepository, 'updateOne')
+        .mockResolvedValue(mockEntity as any);
+
+      const createEscrowCompletionTrackingMock = jest
+        .spyOn(
+          escrowCompletionTrackingService,
+          'createEscrowCompletionTracking',
+        )
+        .mockImplementation(() => {
+          throw new DatabaseError(
+            'Duplicate entry error',
+            PostgresErrorCodes.Duplicated,
+          );
+        });
+
+      await webhookService.processPendingIncomingWebhooks();
+
+      expect(createEscrowCompletionTrackingMock).toHaveBeenCalled();
+      expect(updateOneMock).not.toHaveBeenCalledWith({
+        id: mockEntity.id,
+        status: EscrowCompletionTrackingStatus.COMPLETED,
+      });
+    });
+  });
+
+  describe('processPendingEscrowCompletion', () => {
+    it('should save results and execute paypouts for all of the pending escrows completion', async () => {
+      const mockEntity = {
+        id: 1,
+        status: EscrowCompletionTrackingStatus.PENDING,
+      };
+      jest
+        .spyOn(escrowCompletionTrackingRepository, 'findByStatus')
+        .mockResolvedValue([mockEntity as any]);
+
+      const updateOneMock = jest
+        .spyOn(escrowCompletionTrackingRepository, 'updateOne')
+        .mockResolvedValue(mockEntity as any);
+
+      await webhookService.processPendingEscrowCompletion();
+
+      expect(updateOneMock).toHaveBeenCalled();
+    });
+
+    it('should handle errors and continue processing other entities', async () => {
+      const mockEntity1 = {
+        id: 1,
+        status: EscrowCompletionTrackingStatus.PENDING,
+      };
+      const mockEntity2 = {
+        id: 2,
+        status: EscrowCompletionTrackingStatus.PENDING,
+      };
+
+      jest
+        .spyOn(escrowCompletionTrackingRepository, 'findByStatus')
+        .mockResolvedValue([mockEntity1 as any, mockEntity2 as any]);
+
+      jest
+        .spyOn(escrowCompletionTrackingRepository, 'updateOne')
+        .mockImplementationOnce(() => {
+          throw new Error('Test error');
+        })
+        .mockResolvedValue(mockEntity2 as any);
+
+      await webhookService.processPendingEscrowCompletion();
+    });
+  });
+
+  describe('processPaidEscrowCompletion', () => {
+    let assessReputationScoresMock: jest.SpyInstance;
+    let createOutgoingWebhookMock: jest.SpyInstance;
+
+    beforeEach(() => {
+      assessReputationScoresMock = jest
+        .spyOn(reputationService, 'assessReputationScores')
+        .mockResolvedValue();
+
+      createOutgoingWebhookMock = jest
+        .spyOn(webhookService, 'createOutgoingWebhook')
+        .mockResolvedValue();
+
+      EscrowClient.build = jest.fn().mockResolvedValue({
+        getStatus: jest.fn().mockResolvedValue(EscrowStatus.Paid),
+        complete: jest.fn().mockResolvedValue(true),
+        getJobLauncherAddress: jest.fn().mockResolvedValue(MOCK_ADDRESS),
+        getExchangeOracleAddress: jest.fn().mockResolvedValue(MOCK_ADDRESS),
+        getRecordingOracleAddress: jest.fn().mockResolvedValue(MOCK_ADDRESS),
+      });
+    });
+
+    it('should assess reputation scores and create outgoing webhook for all of the paid escrows completion', async () => {
+      const mockEntity = { id: 1, status: EscrowCompletionTrackingStatus.PAID };
+
+      jest
+        .spyOn(escrowCompletionTrackingRepository, 'findByStatus')
+        .mockResolvedValue([mockEntity as any]);
+
+      const updateOneMock = jest
+        .spyOn(escrowCompletionTrackingRepository, 'updateOne')
+        .mockResolvedValue(mockEntity as any);
+
+      await webhookService.processPaidEscrowCompletion();
+
+      expect(updateOneMock).toHaveBeenCalledWith({
+        id: mockEntity.id,
+        status: EscrowCompletionTrackingStatus.COMPLETED,
+      });
+      expect(assessReputationScoresMock).toHaveBeenCalledTimes(1);
+    });
+
+    it('should handle errors during entity processing without skipping remaining entities', async () => {
+      const mockEntity1 = {
+        id: 1,
+        status: EscrowCompletionTrackingStatus.PAID,
+      };
+      const mockEntity2 = {
+        id: 2,
+        status: EscrowCompletionTrackingStatus.PAID,
+      };
+
+      jest
+        .spyOn(escrowCompletionTrackingRepository, 'findByStatus')
+        .mockResolvedValue([mockEntity1 as any, mockEntity2 as any]);
+
+      const updateOneMock = jest
+        .spyOn(escrowCompletionTrackingRepository, 'updateOne')
+        .mockImplementationOnce(() => {
+          throw new Error('Test error');
+        })
+        .mockResolvedValueOnce(mockEntity2 as any);
+
+      await webhookService.processPaidEscrowCompletion();
+
+      expect(updateOneMock).toHaveBeenCalledTimes(3);
+      expect(assessReputationScoresMock).toHaveBeenCalledTimes(2);
+    });
+
+    it('should retry if createOutgoingWebhook throws a DatabaseError with Duplicated error code', async () => {
+      const mockEntity = { id: 1, status: EscrowCompletionTrackingStatus.PAID };
+
+      jest
+        .spyOn(escrowCompletionTrackingRepository, 'findByStatus')
+        .mockResolvedValue([mockEntity as any]);
+
+      const updateOneMock = jest
+        .spyOn(escrowCompletionTrackingRepository, 'updateOne')
+        .mockResolvedValue(mockEntity as any);
+
+      createOutgoingWebhookMock = jest
+        .spyOn(webhookService, 'createOutgoingWebhook')
+        .mockImplementation(() => {
+          throw new DatabaseError(
+            'Duplicate entry error',
+            PostgresErrorCodes.Duplicated,
+          );
+        });
+
+      await webhookService.processPaidEscrowCompletion();
+
+      expect(createOutgoingWebhookMock).toHaveBeenCalled();
+      expect(updateOneMock).not.toHaveBeenCalledWith({
+        id: mockEntity.id,
+        status: EscrowCompletionTrackingStatus.COMPLETED,
+      });
+    });
+  });
+
+  describe('processPendingOutgoingWebhooks', () => {
+    let sendWebhookMock: any;
+    let webhookEntity1: Partial<WebhookOutgoingEntity>,
+      webhookEntity2: Partial<WebhookOutgoingEntity>;
+
+    beforeEach(() => {
+      webhookEntity1 = {
+        id: 1,
+        payload: {
+          chainId: ChainId.LOCALHOST,
+          escrowAddress: MOCK_ADDRESS,
+          eventType: EventType.ESCROW_COMPLETED,
+        },
+        hash: MOCK_FILE_HASH,
+        url: MOCK_FILE_URL,
+        status: WebhookOutgoingStatus.PENDING,
+        waitUntil: new Date(),
+        retriesCount: 0,
+      };
+
+      webhookEntity2 = {
+        id: 2,
+        payload: {
+          chainId: ChainId.LOCALHOST,
+          escrowAddress: MOCK_ADDRESS,
+          eventType: EventType.ESCROW_COMPLETED,
+        },
+        hash: MOCK_FILE_HASH,
+        url: MOCK_FILE_URL,
+        status: WebhookOutgoingStatus.PENDING,
+        waitUntil: new Date(),
+        retriesCount: 0,
+      };
+
+      jest
+        .spyOn(webhookOutgoingRepository, 'findByStatus')
+        .mockResolvedValue([webhookEntity1 as any, webhookEntity2 as any]);
+
+      sendWebhookMock = jest.spyOn(webhookService as any, 'sendWebhook');
+      sendWebhookMock.mockResolvedValue();
+    });
+
+    afterEach(() => {
+      jest.restoreAllMocks();
+    });
+
+    it('should increase retriesCount by 1 if sending webhook fails', async () => {
+      sendWebhookMock.mockRejectedValueOnce(new Error());
+      await webhookService.processPendingOutgoingWebhooks();
+
+      expect(webhookOutgoingRepository.updateOne).toHaveBeenCalled();
+      expect(webhookEntity1.status).toBe(WebhookOutgoingStatus.PENDING);
+      expect(webhookEntity1.retriesCount).toBe(1);
+      expect(webhookEntity1.waitUntil).toBeInstanceOf(Date);
+    });
+
+    it('should mark webhook as failed if retriesCount exceeds threshold', async () => {
+      sendWebhookMock.mockRejectedValueOnce(new Error());
+
+      webhookEntity1.retriesCount = MOCK_MAX_RETRY_COUNT;
+
+      await webhookService.processPendingOutgoingWebhooks();
+
+      expect(webhookOutgoingRepository.updateOne).toHaveBeenCalled();
+      expect(webhookEntity1.status).toBe(WebhookOutgoingStatus.FAILED);
     });
   });
 });
